feat: add functions for plan upgrade , withdraw

+++ multilevel/contracts/src/state/plan.rs
@@ -6,8 +6,8 @@ pub struct Plan {
     pub investment_required: u64,
     pub validity_days: u32, // no of days plan is valid (daily reward receivable days)
     pub daily_reward: u64,
-    pub max_level: u64, // max level from which referral reward is receivable
-    pub direct_referral_percentage: u32,
-    pub active_referral_percentage: u32, //level 2 - 6 ,
-    pub deep_referral_percentage: u32,   //level 7 - max level
+    pub max_level: u8, // max level from which referral reward is receivable
+    pub direct_referral_percentage: u8,
+    pub active_referral_percentage: u8, //level 2 - 6 ,
+    pub deep_referral_percentage: u8,   //level 7 - max level
 }
